+++ src/main/java/com/marginallyclever/ro3/Registry.java
@@ -5,6 +5,9 @@ import com.marginallyclever.ro3.mesh.MeshFactory;
 import com.marginallyclever.ro3.node.nodes.*;
 import com.marginallyclever.ro3.node.Node;
 import com.marginallyclever.ro3.node.nodes.marlinrobotarm.MarlinRobotArm;
+import com.marginallyclever.ro3.node.nodes.pose.Limb;
+import com.marginallyclever.ro3.node.nodes.pose.LookAt;
+import com.marginallyclever.ro3.node.nodes.pose.MeshInstance;
 import com.marginallyclever.ro3.texture.TextureFactory;
 
 import javax.swing.event.EventListenerList;
@@ -34,6 +37,7 @@ public class Registry {
         Factory.Category<Node> nodule = nodeFactory.getRoot().add("Node", Node::new);
         nodule.add("DHParameter", DHParameter::new);
         nodule.add("HingeJoint", HingeJoint::new);
+        nodule.add("LimbSolver", LimbSolver::new);
         nodule.add("MarlinRobotArm", MarlinRobotArm::new);
         nodule.add("Material", Material::new);
         nodule.add("MeshInstance", MeshInstance::new);
@@ -41,7 +45,7 @@ public class Registry {
         Factory.Category<Node> pose = nodule.add("Pose", Pose::new);
             pose.add("Camera", Camera::new);
             pose.add("LookAt", LookAt::new);
-
+            pose.add("Limb", Limb::new);
         reset();
     }
