+++ src/main/java/com/marginallyclever/ro3/node/nodes/marlinrobotarm/MarlinRobotArm.java
@@ -2,16 +2,15 @@ package com.marginallyclever.ro3.node.nodes.marlinrobotarm;
 
 import com.marginallyclever.convenience.helpers.MatrixHelper;
 import com.marginallyclever.convenience.helpers.StringHelper;
-import com.marginallyclever.convenience.swing.Dial;
 import com.marginallyclever.convenience.swing.NumberFormatHelper;
 import com.marginallyclever.ro3.apps.nodedetailview.CollapsiblePanel;
 import com.marginallyclever.ro3.node.Node;
 import com.marginallyclever.ro3.node.NodePath;
 import com.marginallyclever.ro3.node.nodes.HingeJoint;
+import com.marginallyclever.ro3.node.nodes.LimbSolver;
 import com.marginallyclever.ro3.node.nodes.Motor;
 import com.marginallyclever.ro3.node.nodes.Pose;
-import com.marginallyclever.ro3.apps.nodeselector.NodeSelector;
-import org.json.JSONArray;
+import com.marginallyclever.ro3.node.nodes.pose.Limb;
 import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,8 +22,6 @@ import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.HierarchyEvent;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
@@ -32,22 +29,17 @@ import java.util.Objects;
  * <p>{@link MarlinRobotArm} converts the state of a robot arm into GCode and back.</p>
  * <p>In order to work it requires references to:</p>
  * <ul>
- *     <li>five or six {@link Motor}s, with names matching those in Marlin;</li>
- *     <li>a {@link Pose} end effector to obtain the inverse kinematic pose.  The end effector should be at the end of
- *     the kinematic chain.</li>
- *     <li>a {@link Pose} target that the end effector will try to match.  It will only do so when the linear velocity
- *     is greater than zero.</li>
+ *     <li>a {@link Limb} of no more than six {@link Motor}s, whose names match those in Marlin;</li>
+ *     <li>a {@link LimbSolver} to calculate the inverse kinematics;</li>
  *     <li>an optional {@link Motor} for the tool on arm.</li>
  * </ul>
  */
 public class MarlinRobotArm extends Node {
     private static final Logger logger = LoggerFactory.getLogger(MarlinRobotArm.class);
-    public static final int MAX_JOINTS = 6;
-    private final List<NodePath<Motor>> motors = new ArrayList<>();
+    public final Limb limb = new Limb();
+    public final LimbSolver solver = new LimbSolver();
     private final NodePath<Motor> gripperMotor = new NodePath<>(this,Motor.class);
-    private final NodePath<Pose> endEffector = new NodePath<>(this,Pose.class);
-    private final NodePath<Pose> target = new NodePath<>(this,Pose.class);
-    private double linearVelocity=0;
+    private double reportInterval=1.0;
 
     public MarlinRobotArm() {
         this("MarlinRobotArm");
@@ -55,25 +47,15 @@ public class MarlinRobotArm extends Node {
 
     public MarlinRobotArm(String name) {
         super(name);
-        for(int i=0;i<MAX_JOINTS;++i) {
-            motors.add(new NodePath<>(this,Motor.class));
-        }
     }
 
     @Override
     public JSONObject toJSON() {
         JSONObject json = super.toJSON();
-        JSONArray jointArray = new JSONArray();
-        json.put("version",1);
-
-        for(var motor : motors) {
-            jointArray.put(motor == null ? JSONObject.NULL : motor.getPath());
-        }
-        json.put("motors",jointArray);
-        if(endEffector.getSubject()!=null) json.put("endEffector",endEffector.getPath());
-        if(target.getSubject()!=null) json.put("target",target.getPath());
+        json.put("version",2);
+        json.put("limb",limb.toJSON());
+        json.put("solver",solver.toJSON());
         if(gripperMotor.getSubject()!=null) json.put("gripperMotor",gripperMotor.getPath());
-        json.put("linearVelocity",linearVelocity);
         return json;
     }
 
@@ -81,52 +63,23 @@ public class MarlinRobotArm extends Node {
     public void fromJSON(JSONObject from) {
         super.fromJSON(from);
         int version = from.has("version") ? from.getInt("version") : 0;
-
-        if(from.has("motors")) {
-            JSONArray motorArray = from.getJSONArray("motors");
-            for(int i=0;i<motorArray.length();++i) {
-                if(motorArray.isNull(i)) {
-                    motors.get(i).setPath(null);
-                } else {
-                    if(version==1) {
-                        motors.get(i).setPath(motorArray.getString(i));
-                    } else if(version==0) {
-                        Motor motor = this.getRootNode().findNodeByID(motorArray.getString(i), Motor.class);
-                        motors.get(i).setRelativePath(this,motor);
-                    }
+        if(version==2) {
+            limb.fromJSON(from.getJSONObject("limb"));
+            solver.fromJSON(from.getJSONObject("solver"));
+            return;
+        } else {
+            limb.fromJSON(from);
+            solver.fromJSON(from);
+            Node root = this.getRootNode();
+            if (from.has("gripperMotor")) {
+                String s = from.getString("gripperMotor");
+                if (version == 1) {
+                    gripperMotor.setPath(s);
+                } else if (version == 0) {
+                    Motor goal = root.findNodeByID(s, Motor.class);
+                    gripperMotor.setRelativePath(this, goal);
                 }
             }
-        }
-        Node root = this.getRootNode();
-        if(from.has("endEffector")) {
-            String s = from.getString("endEffector");
-            if(version==1) {
-                endEffector.setPath(s);
-            } else if(version==0) {
-                Pose goal = root.findNodeByID(s,Pose.class);
-                endEffector.setRelativePath(this,goal);
-            }
-        }
-        if(from.has("target")) {
-            String s = from.getString("target");
-            if(version==1) {
-                target.setPath(s);
-            } else if(version==0) {
-                Pose goal = root.findNodeByID(s,Pose.class);
-                target.setRelativePath(this,goal);
-            }
-        }
-        if(from.has("gripperMotor")) {
-            String s = from.getString("gripperMotor");
-            if(version==1) {
-                gripperMotor.setPath(s);
-            } else if(version==0) {
-                Motor goal = root.findNodeByID(s,Motor.class);
-                gripperMotor.setRelativePath(this,goal);
-            }
-        }
-        if(from.has("linearVelocity")) {
-            linearVelocity = from.getDouble("linearVelocity");
         }
     }
 
@@ -143,13 +96,7 @@ public class MarlinRobotArm extends Node {
 
         gbc.gridx=0;
         gbc.gridy=0;
-        gbc.gridwidth=2;
-        pane.add(createVelocitySlider(),gbc);
-
-        gbc.gridy++;
         gbc.gridwidth=1;
-
-        addMoveTargetToEndEffector(pane,gbc);
 
         JButton M114 = new JButton("M114");
         M114.addActionListener(e-> sendGCode("M114"));
@@ -157,9 +104,6 @@ public class MarlinRobotArm extends Node {
 
         gbc.gridwidth=1;
         gbc.gridy++;
-
-        addNodeSelector(pane, "End Effector", endEffector, Pose.class, gbc);
-        addNodeSelector(pane, "Target", target, Pose.class, gbc);
         addNodeSelector(pane, "Gripper motor", gripperMotor, Motor.class, gbc);
 
         gbc.gridx=0;
@@ -169,33 +113,12 @@ public class MarlinRobotArm extends Node {
         pane.add(getSender(),gbc);
         gbc.gridy++;
         pane.add(createReportInterval(),gbc);
-        gbc.gridy++;
-        pane.add(createEasyFK(),gbc);
-        gbc.gridy++;
 
-        gbc.gridwidth=2;
-        pane.add(addMotorPanel(),gbc);
+        solver.getComponents(list);
+        limb.getComponents(list);
 
         super.getComponents(list);
     }
-
-    private JComponent createVelocitySlider() {
-        JPanel container = new JPanel(new BorderLayout());
-        // add a slider to control linear velocity
-        JSlider slider = new JSlider(0,20,(int)linearVelocity);
-        slider.addChangeListener(e-> linearVelocity = slider.getValue());
-
-        // Make the slider fill the available horizontal space
-        slider.setMaximumSize(new Dimension(Integer.MAX_VALUE, slider.getPreferredSize().height));
-        slider.setMinimumSize(new Dimension(50, slider.getPreferredSize().height));
-
-        container.add(new JLabel("Linear Vel"), BorderLayout.LINE_START);
-        container.add(slider, BorderLayout.CENTER); // Add slider to the center of the container
-
-        return container;
-    }
-
-    private double reportInterval=1.0;
 
     private JComponent createReportInterval() {
         var containerPanel = new CollapsiblePanel("Report");
@@ -277,96 +200,6 @@ public class MarlinRobotArm extends Node {
         reportInterval = Math.max(0.1,seconds);
     }
 
-
-    private <T extends Node> void addNodeSelector(JPanel pane, String label, NodePath<T> nodePath, Class<T> clazz, GridBagConstraints gbc) {
-        NodeSelector<T> selector = new NodeSelector<>(clazz, nodePath.getSubject());
-        selector.addPropertyChangeListener("subject", (e) -> nodePath.setRelativePath(this, (T) e.getNewValue()));
-        addLabelAndComponent(pane, label, selector, gbc);
-    }
-
-    private void addMoveTargetToEndEffector(JPanel pane,GridBagConstraints gbc) {
-        // move target to end effector
-        JButton targetToEE = new JButton(new AbstractAction() {
-            {
-                putValue(Action.NAME,"Move");
-                putValue(Action.SHORT_DESCRIPTION,"Move the Target Pose to the End Effector.");
-                putValue(Action.SMALL_ICON,new ImageIcon(Objects.requireNonNull(getClass().getResource(
-                        "/com/marginallyclever/ro3/apps/shared/icons8-move-16.png"))));
-            }
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                setTargetToEndEffector();
-            }
-        });
-        addLabelAndComponent(pane, "Target to EE", targetToEE,gbc);
-    }
-
-    private void setTargetToEndEffector() {
-        if(target.getSubject()!=null && endEffector.getSubject()!=null) {
-            target.getSubject().setWorld(endEffector.getSubject().getWorld());
-        }
-    }
-
-    private JComponent addMotorPanel() {
-        var containerPanel = new CollapsiblePanel("Motors");
-        var outerPanel = containerPanel.getContentPane();
-        outerPanel.setLayout(new GridBagLayout());
-
-        GridBagConstraints gbc = new GridBagConstraints();
-        gbc.weightx = 1.0;
-        gbc.weighty = 1.0;
-        gbc.gridx=0;
-        gbc.gridy=0;
-        gbc.fill = GridBagConstraints.BOTH;
-
-        // add a selector for each motor
-        var motorSelector = new NodeSelector[MAX_JOINTS];
-        for(int i=0;i<MAX_JOINTS;++i) {
-            motorSelector[i] = new NodeSelector<>(Motor.class, motors.get(i).getSubject());
-            int j = i;
-            motorSelector[i].addPropertyChangeListener("subject",(e)-> {
-                motors.get(j).setRelativePath(this,(Motor)e.getNewValue());
-            });
-            addLabelAndComponent(outerPanel, "Motor "+i, motorSelector[i],gbc);
-        }
-        return containerPanel;
-    }
-
-    private JComponent createEasyFK() {
-        var containerPanel = new CollapsiblePanel("Forward Kinematics");
-        var outerPanel = containerPanel.getContentPane();
-        outerPanel.setLayout(new GridLayout(0,3));
-
-        int count=0;
-        for(int i=0;i<getNumJoints();++i) {
-            final Motor motor = motors.get(i).getSubject();
-            if(motor!=null) {
-                JPanel innerPanel = new JPanel(new BorderLayout());
-                Dial dial = new Dial();
-                dial.addActionListener(e -> {
-                    motor.getHinge().setAngle(dial.getValue());
-                });
-                // TODO subscribe to motor.getAxle().getAngle(), then dial.setValue() without triggering an action event.
-
-                JLabel label = new JLabel(motor.getName());
-                label.setLabelFor(dial);
-                label.setHorizontalAlignment(SwingConstants.CENTER);
-                innerPanel.add(label,BorderLayout.PAGE_START);
-                innerPanel.add(dial,BorderLayout.CENTER);
-                dial.setPreferredSize(new Dimension(80,80));
-
-                outerPanel.add(innerPanel);
-                count++;
-            }
-        }
-        count = 3-(count%3);
-        for(int i=0;i<count;++i) {
-            outerPanel.add(new JPanel());
-        }
-
-        return containerPanel;
-    }
-
     // Add a text field that will be sent to the robot arm.
     private JPanel getSender() {
         JPanel inputPanel = new JPanel(new BorderLayout());
@@ -406,7 +239,7 @@ public class MarlinRobotArm extends Node {
 
     private String getMotorsAndFeedrateAsString() {
         StringBuilder sb = new StringBuilder();
-        for(NodePath<Motor> paths : this.motors) {
+        for(NodePath<Motor> paths : limb.getMotors()) {
             Motor motor = paths.getSubject();
             if(motor!=null && motor.hasAxle()) {
                 sb.append(" ")
@@ -422,7 +255,7 @@ public class MarlinRobotArm extends Node {
                     .append(StringHelper.formatDouble(gripperMotor.getHinge().getAngle()));
         }
         // feedrate
-        sb.append(" F").append(StringHelper.formatDouble(linearVelocity));
+        sb.append(" F").append(StringHelper.formatDouble(solver.getLinearVelocity()));
         return sb.toString();
     }
 
@@ -444,7 +277,7 @@ public class MarlinRobotArm extends Node {
         } else if(gcode.equals("ik")) {
             fireMarlinMessage(getEndEffectorIK());
         } else if(gcode.equals("aj")) {
-            ApproximateJacobianFiniteDifferences jacobian = new ApproximateJacobianFiniteDifferences(this);
+            ApproximateJacobianFiniteDifferences jacobian = new ApproximateJacobianFiniteDifferences(limb);
             fireMarlinMessage( "Ok: "+jacobian );
             return;
         } else if(gcode.startsWith("G1")) {
@@ -463,7 +296,7 @@ public class MarlinRobotArm extends Node {
     private String parseG0(String gcode) {
         String [] parts = gcode.split("\\s+");
         try {
-            for (NodePath<Motor> paths : this.motors) {
+            for (NodePath<Motor> paths : limb.getMotors()) {
                 Motor motor = paths.getSubject();
                 if (motor != null && motor.hasAxle()) {
                     for (String p : parts) {
@@ -499,24 +332,24 @@ public class MarlinRobotArm extends Node {
      * <p>G1 Linear move.</p>
      * <p>Parse gcode for names and values, then set the new target position.  Names are XYZ for linear, UVW for
      * angular. Angular values should be in degrees.</p>
-     * <p>Movement will occur on {@link #update(double)} provided the {@link #linearVelocity} and the update time are
-     * greater than zero.</p>
+     * <p>Movement will occur on {@link #update(double)} provided the {@link LimbSolver} linear velocity and the update
+     * time are greater than zero.</p>
      * @param gcode GCode command
      * @return response from robot arm
      */
     private String parseG1(String gcode) {
-        if(target.getSubject()==null) {
+        if(solver.getTarget()==null) {
             logger.warn("no target");
             return "Error: no target";
         }
-        if(endEffector.getSubject()==null) {
+        if(limb.getEndEffector()==null) {
             logger.warn("no end effector");
             return "Error: no end effector";
         }
         String [] parts = gcode.split("\\s+");
-        double [] cartesian = getCartesianFromWorld(endEffector.getSubject().getWorld());
+        double [] cartesian = getCartesianFromWorld(limb.getEndEffector().getWorld());
         for(String p : parts) {
-            if(p.startsWith("F")) setLinearVelocity(Double.parseDouble(p.substring(1)));
+            if(p.startsWith("F")) solver.setLinearVelocity(Double.parseDouble(p.substring(1)));
             if(p.startsWith("X")) cartesian[0] = Double.parseDouble(p.substring(1));
             else if(p.startsWith("Y")) cartesian[1] = Double.parseDouble(p.substring(1));
             else if(p.startsWith("Z")) cartesian[2] = Double.parseDouble(p.substring(1));
@@ -526,18 +359,18 @@ public class MarlinRobotArm extends Node {
             else logger.warn("unknown G1 command: "+p);
         }
         // set the target position relative to the base of the robot arm
-        target.getSubject().setLocal(getReverseCartesianFromWorld(cartesian));
+        solver.getTarget().setLocal(getReverseCartesianFromWorld(cartesian));
         return "Ok";
     }
 
     private String getEndEffectorIK() {
-        if(endEffector.getSubject()==null) {
+        if(limb.getEndEffector()==null) {
             return ( "Error: no end effector" );
         }
-        double [] cartesian = getCartesianFromWorld(endEffector.getSubject().getWorld());
+        double [] cartesian = getCartesianFromWorld(limb.getEndEffector().getWorld());
         int i=0;
         String response = "G1"
-                +" F"+StringHelper.formatDouble(linearVelocity)
+                +" F"+StringHelper.formatDouble(solver.getLinearVelocity())
                 +" X"+StringHelper.formatDouble(cartesian[i++])
                 +" Y"+StringHelper.formatDouble(cartesian[i++])
                 +" Z"+StringHelper.formatDouble(cartesian[i++])
@@ -578,20 +411,16 @@ public class MarlinRobotArm extends Node {
     public int getNumJoints() {
         // count the number of motors such that motors.get(i).getSubject()!=null
         int count=0;
-        for(NodePath<Motor> paths : motors) {
+        for(NodePath<Motor> paths : limb.getMotors()) {
             if(paths.getSubject()!=null) count++;
         }
         return count;
     }
 
-    public Motor getJoint(int i) {
-        return motors.get(i).getSubject();
-    }
-
     public double[] getAllJointAngles() {
         double[] result = new double[getNumJoints()];
         int i=0;
-        for(NodePath<Motor> paths : motors) {
+        for(NodePath<Motor> paths : limb.getMotors()) {
             Motor motor = paths.getSubject();
             if(motor!=null) {
                 result[i++] = motor.getHinge().getAngle();
@@ -606,7 +435,7 @@ public class MarlinRobotArm extends Node {
             return;
         }
         int i=0;
-        for(NodePath<Motor> paths : motors) {
+        for(NodePath<Motor> paths : limb.getMotors()) {
             Motor motor = paths.getSubject();
             if(motor!=null) {
                 HingeJoint axle = motor.getHinge();
@@ -624,7 +453,7 @@ public class MarlinRobotArm extends Node {
             return;
         }
         int i=0;
-        for(NodePath<Motor> paths : motors) {
+        for(NodePath<Motor> paths : limb.getMotors()) {
             Motor motor = paths.getSubject();
             if(motor!=null) {
                 HingeJoint axle = motor.getHinge();
@@ -639,124 +468,14 @@ public class MarlinRobotArm extends Node {
      * @return the end effector pose or null if not set.
      */
     public Pose getEndEffector() {
-        return endEffector.getSubject() == null ? null : endEffector.getSubject();
+        return limb.getEndEffector();
     }
 
     /**
      * @return the target pose or null if not set.
      */
     public Pose getTarget() {
-        return target.getSubject() == null ? null : target.getSubject();
-    }
-
-    public void setTarget(Pose target) {
-        this.target.setRelativePath(this,target);
-    }
-
-    @Override
-    public void update(double dt) {
-        super.update(dt);
-        if(dt==0) return;
-        moveTowardsTarget(dt);
-    }
-
-    private void moveTowardsTarget(double dt) {
-        if(endEffector.getSubject()==null || target.getSubject()==null || linearVelocity<0.0001) {
-            double[] jointAnglesOriginal = getAllJointAngles();
-            Arrays.fill(jointAnglesOriginal, 0);
-            setAllJointVelocities(jointAnglesOriginal);
-            return;
-        }
-        double[] cartesianVelocity = MatrixHelper.getCartesianBetweenTwoMatrices(
-                endEffector.getSubject().getWorld(),
-                target.getSubject().getWorld());
-        scaleVectorToMagnitude(cartesianVelocity,linearVelocity);
-        moveEndEffectorInCartesianDirection(cartesianVelocity);
-    }
-
-    /**
-     * Make sure the given vector's length does not exceed maxLen.  It can be less than the given magnitude.
-     * Store the results in the original array.
-     * @param vector the vector to cap
-     * @param maxLen the max length of the vector.
-     */
-    public static void scaleVectorToMagnitude(double[] vector, double maxLen) {
-        // get the length of the vector
-        double len = 0;
-        for (double v : vector) {
-            len += v * v;
-        }
-
-        len = Math.sqrt(len);
-        if(maxLen>len) maxLen=len;
-
-        // scale the vector
-        double scale = len==0? 0 : maxLen / len;  // catch len==0
-        for(int i=0;i<vector.length;i++) {
-            vector[i] *= scale;
-        }
-    }
-
-    /**
-     * Attempts to move the robot arm such that the end effector travels in the direction of the cartesian velocity.
-     * @param cartesianVelocity three linear forces (mm) and three angular forces (degrees).
-     * @throws RuntimeException if the robot cannot be moved in the direction of the cartesian force.
-     */
-    public void moveEndEffectorInCartesianDirection(double[] cartesianVelocity) {
-        // is it a tiny move?
-        double sum = sumCartesianVelocityComponents(cartesianVelocity);
-        if(sum<0.0001) return;
-        if(sum <= 1) {
-            setMotorVelocitiesFromCartesianVelocity(cartesianVelocity);
-            return;
-        }
-
-        // set motor velocities.
-        setMotorVelocitiesFromCartesianVelocity(cartesianVelocity);
-    }
-
-    /**
-     * <p>Attempts to move the robot arm such that the end effector travels in the cartesian direction.  This is
-     * achieved by setting the velocity of the motors.</p>
-     * @param cartesianVelocity three linear forces (mm) and three angular forces (degrees).
-     * @throws RuntimeException if the robot cannot be moved in the direction of the cartesian force.
-     */
-    private void setMotorVelocitiesFromCartesianVelocity(double[] cartesianVelocity) {
-        ApproximateJacobian aj = getJacobian();
-        try {
-            double[] jointVelocity = aj.getJointFromCartesian(cartesianVelocity);  // uses inverse jacobian
-            if(impossibleVelocity(jointVelocity)) return;  // TODO: throw exception instead?
-            setAllJointVelocities(jointVelocity);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private ApproximateJacobian getJacobian() {
-        // option 1, use finite differences
-        return new ApproximateJacobianFiniteDifferences(this);
-        // option 2, use screw theory
-        //ApproximateJacobian aj = new ApproximateJacobianScrewTheory(robotComponent);
-    }
-
-    /**
-     * @param jointVelocity the joint velocity to check
-     * @return true if the given joint velocity is impossible.
-     */
-    private boolean impossibleVelocity(double[] jointVelocity) {
-        double maxV = 100; // RPM*60 TODO: get from robot per joint
-        for(double v : jointVelocity) {
-            if(Double.isNaN(v) || Math.abs(v) > maxV) return true;
-        }
-        return false;
-    }
-
-    private double sumCartesianVelocityComponents(double [] cartesianVelocity) {
-        double sum = 0;
-        for (double v : cartesianVelocity) {
-            sum += Math.abs(v);
-        }
-        return sum;
+        return solver.getTarget();
     }
 
     public void addMarlinListener(MarlinListener editorPanel) {
@@ -773,14 +492,5 @@ public class MarlinRobotArm extends Node {
         for(MarlinListener listener : listeners.getListeners(MarlinListener.class)) {
             listener.messageFromMarlin(message);
         }
-    }
-
-    public double getLinearVelocity() {
-        return linearVelocity;
-    }
-
-    public void setLinearVelocity(double linearVelocity) {
-        if(linearVelocity<0) throw new IllegalArgumentException("linearVelocity must be >= 0");
-        this.linearVelocity = linearVelocity;
     }
 }
