+++ src/test/java/com/marginallyclever/ro3/node/nodes/marlinrobotarm/ApproximateJacobianTest.java
@@ -2,6 +2,7 @@ package com.marginallyclever.ro3.node.nodes.marlinrobotarm;
 
 import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.apps.actions.LoadScene;
+import com.marginallyclever.ro3.node.nodes.pose.Limb;
 import org.junit.jupiter.api.*;
 
 import java.io.File;
@@ -19,13 +20,13 @@ public class ApproximateJacobianTest {
         Registry.start();
     }
 
-    private MarlinRobotArm build6AxisArm() throws Exception {
+    private Limb build6AxisArm() throws Exception {
         // TODO load a robot from a file.
         var load = new LoadScene(null,null);
         // find file {project root}/src/test/resources/com/marginallyclever/ro3/apps/node/nodes/marlinrobotarm/Sixi3-5.RO
         File file = new File("src/test/resources/com/marginallyclever/ro3/apps/node/nodes/marlinrobotarm/Sixi3-5.RO");
         load.commitLoad(file);
-        return (MarlinRobotArm) Registry.getScene().get("./Sixi3/MarlinRobotArm");
+        return (Limb) Registry.getScene().get("./Sixi3/MarlinRobotArm");
     }
 
     /**
@@ -35,7 +36,7 @@ public class ApproximateJacobianTest {
     @Test
     @Disabled
     public void compare() throws Exception {
-        MarlinRobotArm robot = build6AxisArm();
+        Limb robot = build6AxisArm();
         ApproximateJacobian finite = new ApproximateJacobianFiniteDifferences(robot);
         ApproximateJacobian screw = new ApproximateJacobianScrewTheory(robot);
